arch: Implement Faux's Intelliplug
As much as I despise shying away from QC's mpdecision, I was faced with a decision. My CPU toggle features piss off mpdecision, which goes hectic in the logs because it doesn't know what to do. Atleast if I use Intelliplug, I can tweak the source to add in my own conditions for when a CPU is disabled. Blame QC being closed source for whatever reason